Creacion de tipos de Usuarios, + algunas funcionalidades

+++ src/models/pedido.cpp
@@ -1,11 +1,12 @@
 #include "pedido.h"
+#include "pedidoitem.h"
 #include <iostream>
 #include <string>
 
 void Pedido::agregarItem(Item* item, int cantidad, const std::string& comentario) {
     items.push_back(new PedidoItem(item, cantidad, comentario));
 }
-int Pedido::getMontoTotal() const {
+double Pedido::getMontoTotal() const {
             int total = 0;
             for(PedidoItem* item : items){
                 total += item->getItem()->getPrecio() *item->getCantidad();
